bot_service/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.0
bitarray==2.7.3
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.1
cytoolz==0.12.1
decorator==5.1.1
Deprecated==1.2.13
discord==2.2.2
discord.py==2.2.2
dnspython==2.3.0
email-validator==1.3.1
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
fastapi==0.101.1
fuzzywuzzy==0.18.0
filelock==3.10.7
frozenlist==1.3.3
google-auth==2.17.1
google-auth-oauthlib==1.0.0
gspread==5.7.2
h11==0.14.0
hexbytes==0.3.0
httplib2==0.22.0
huggingface-hub==0.13.3
idna==3.4
joblib==1.2.0
jsonschema==4.17.3
lru-dict==1.1.8
multidict==6.0.4
nltk==3.8.1
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.2
packaging==23.0
pandas==1.5.3
parsimonious==0.9.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
PyGithub==1.58.1
PyJWT==2.6.0
pymongo==4.3.3
PyMuPDF==1.21.1
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
Requests==2.31.0
requests-oauthlib==1.3.1
rlp==3.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
starlette==0.26.1
tiktoken==0.3.3
tokenizers==0.13.2
toolz==0.12.0
tqdm==4.65.0
transformers==4.27.4
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
validators==0.20.0
web3==6.0.0
websockets==11.0
wrapt==1.15.0
yarl==1.8.2